Prep for better history support.
Distinction between continuation and callstate allows me to determine if
two different entities arrived at the same state and expression.  For
example an if statement with an else branch that neither branch has side
effect to the scope chain then after that they arrive in the same state,
but parent continuation is different.